move utility functions
