+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/database_logic.py
@@ -27,7 +27,13 @@ from stac_fastapi.sfeos_helpers.database import (
     create_index_templates_shared,
     delete_item_index_shared,
     get_queryables_mapping_shared,
+    index_alias_by_collection_id,
+    index_by_collection_id,
+    indices,
+    mk_actions,
+    mk_item_id,
     populate_sort_shared,
+    validate_refresh,
 )
 from stac_fastapi.sfeos_helpers.mappings import (
     AGGREGATION_MAPPING,
@@ -36,14 +42,6 @@ from stac_fastapi.sfeos_helpers.mappings import (
     ITEM_INDICES,
     ITEMS_INDEX_PREFIX,
     Geometry,
-)
-from stac_fastapi.sfeos_helpers.utilities import (
-    index_alias_by_collection_id,
-    index_by_collection_id,
-    indices,
-    mk_actions,
-    mk_item_id,
-    validate_refresh,
 )
 from stac_fastapi.types.errors import ConflictError, NotFoundError
 from stac_fastapi.types.stac import Collection, Item
