+++ stac_fastapi/tests/database/test_database.py
@@ -3,12 +3,12 @@ import uuid
 import pytest
 from stac_pydantic import api
 
+from stac_fastapi.sfeos_helpers.database import index_alias_by_collection_id
 from stac_fastapi.sfeos_helpers.mappings import (
     COLLECTIONS_INDEX,
     ES_COLLECTIONS_MAPPINGS,
     ES_ITEMS_MAPPINGS,
 )
-from stac_fastapi.sfeos_helpers.utilities import index_alias_by_collection_id
 
 from ..conftest import MockRequest, database
